pyproject.toml
[project]
name = "adkagente"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.4.2",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.0",
]
